+++ grid_map_core/include/grid_map_core/iterators/LineIterator.hpp
@@ -27,18 +27,18 @@ public:
   /*!
    * Constructor.
    * @param gridMap the grid map to iterate on.
-   * @param start the starting index of the line.
-   * @param end the ending index of the line.
+   * @param start the starting point of the line.
+   * @param end the ending point of the line.
    */
-  LineIterator(const grid_map::GridMap& gridMap, const Eigen::Array2i& start, const Eigen::Array2i& end);
+  LineIterator(const grid_map::GridMap& gridMap, const Position& start, const Position& end);
 
   /*!
    * Constructor.
    * @param gridMap the grid map to iterate on.
-   * @param start the starting point of the line.
-   * @param end the ending point of the line.
+   * @param start the starting index of the line.
+   * @param end the ending index of the line.
    */
-  LineIterator(const grid_map::GridMap& gridMap, const Eigen::Vector2d& start, const Eigen::Vector2d& end);
+  LineIterator(const grid_map::GridMap& gridMap, const Index& start, const Index& end);
 
   /*!
    * Assignment operator.
@@ -57,7 +57,7 @@ public:
    * Dereference the iterator with const.
    * @return the value to which the iterator is pointing.
    */
-  const Eigen::Array2i& operator *() const;
+  const Index& operator *() const;
 
   /*!
    * Increase the iterator to the next element.
@@ -73,17 +73,40 @@ public:
 
 private:
 
-  // TODO
-  void initializeParameters();
+
+  /*!
+   * Construct function.
+   * @param gridMap the grid map to iterate on.
+   * @param start the starting index of the line.
+   * @param end the ending index of the line.
+   * @return true if successful, false otherwise.
+   */
+  bool initialize(const grid_map::GridMap& gridMap, const Index& start, const Index& end);
+
+  /*!
+   * Computes the parameters requires for the line drawing algorithm.
+   */
+  void initializeIterationParameters();
+
+  /*!
+   * Finds the index of a position on a line within the limits of the map.
+   * @param[in] gridMap the grid map that defines the map boundaries.
+   * @param[in] start the position that will be limited to the map range.
+   * @param[in] end the ending position of the line.
+   * @param[out] index the index of the moved start position.
+   * @return true if successful, false otherwise.
+   */
+  bool getIndexLimitedToMapRange(const grid_map::GridMap& gridMap, const Position& start,
+                                 const Position& end, Index& index);
 
   //! Current index.
-  Eigen::Array2i index_;
+  Index index_;
 
   //! Starting index of the line.
-  Eigen::Array2i start_;
+  Index start_;
 
   //! Ending index of the line.
-  Eigen::Array2i end_;
+  Index end_;
 
   //! Current cell number.
   unsigned int iCell_;
@@ -92,15 +115,15 @@ private:
   unsigned int nCells_;
 
   //! Helper variables for Bresenham Line Drawing algorithm.
-  Eigen::Array2i increment1_, increment2_;
+  Size increment1_, increment2_;
   int denominator_, numerator_, numeratorAdd_;
 
   //! Map information needed to get position from iterator.
-  Eigen::Array2d mapLength_;
-  Eigen::Vector2d mapPosition_;
+  Length mapLength_;
+  Position mapPosition_;
   double resolution_;
-  Eigen::Array2i bufferSize_;
-  Eigen::Array2i bufferStartIndex_;
+  Size bufferSize_;
+  Index bufferStartIndex_;
 };
 
 } /* namespace */