+++ grid_map/CHANGELOG.rst
@@ -2,6 +2,9 @@
 Changelog for package grid_map
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
+Forthcoming
+-----------
+
 1.4.0 (2016-08-22)
 ------------------
 * Added new package grid_map_rviz_plugin.